sdk/python/pyproject.toml
# SPDX-FileCopyrightText: 2025 NodeCube d.o.o. and contributors
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "securefabric-sdk"
version = "0.1.0"
description = "Python SDK for SecureFabric secure messaging"
authors = [{name = "NodeCube Team", email = "legal@nodecube.io"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.9"
keywords = ["securefabric", "messaging", "grpc", "encryption", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "pynacl>=1.5.0",
    "blake3>=0.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "grpc-stubs>=1.53.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["securefabric*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
